requirements.txt
# Core web framework
flask==3.1.2
gunicorn==23.0.0

# Image processing
opencv-python==4.12.0.88
pillow==11.3.0
pillow_heif==1.1.1
matplotlib==3.9.4
numpy==2.1.3

# Deep learning (CPU wheels only) Python 3.11 compatible
# PyTorch 2.4.x ships CPU wheels on PyPI; no +cpu suffix needed
torch==2.4.1
torchvision==0.19.1
facenet-pytorch==2.5.3

# No extra index needed for CPU wheels on PyPI
